Command my-window-at-top using C-c i
Add a new command, my-window-at-top. This command opens a new window of
the same buffer but jumps to the top of the file. This can be useful,
for instance, to quickly add a new import. In this window, closing the
window has a temporary keybinding as C-c C-c (which required adding a
minor mode).

The command is bound to C-c i.